requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.0.2
flask-cors==5.0.1
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pillow==10.2.0
scikit-image==0.22.0
scipy==1.15.2
tifffile==2025.3.30
Werkzeug==3.1.3